requirements.txt
aiohttp
aiohttp_retry
apscheduler
beautifulsoup4
clean_links
feedparser
html_telegraph_poster
lxml[html_clean]
pymongo
python-decouple
python-ffmpeg
pytz
urlextract
w3lib
git+https://github.com/Nubuki-all/neonize@patch-1
git+https://github.com/Nubuki-all/Real-ESRGAN
#Snyk pins 
certifi>=2023.7.22 # not directly required, pinned by Snyk to avoid a vulnerability
dnspython>=2.6.0rc1 # not directly required, pinned by Snyk to avoid a vulnerability
requests>=2.32.0 # not directly required, pinned by Snyk to avoid a vulnerability
setuptools>=65.5.1 # not directly required, pinned by Snyk to avoid a vulnerability
